requirements.txt
streamlit
yfinance
pandas
openpyxl
python-dateutil